# openapi.yaml
...
openapi: "3.0.3"
info:
  title: "identuum-idp-oss API"
  version: "identuum-idp-oss-api-docgen 0.2.0-p3"
  description: "Machine-generated OpenAPI 3 specification for the identuum-idp-oss API surface. Generated by tools/api-docgen from // docgen: source annotations. Response bodies without a known schema record are described but not field-modeled; see the referenced Go symbol in identuum-idp-oss for the authoritative shape."
paths:
...
  "/api/v1/organizations/{id}/service-accounts":
    get:
      operationId: "identuum-idp-oss.service-accounts.list_service_accounts"
      summary: "List service accounts in an organization (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "List service accounts in an organization (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
    post:
      operationId: "identuum-idp-oss.service-accounts.create_service_account"
      summary: "Create a service account in an organization (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "201":
          description: "Create a service account in an organization (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
...
  "/api/v1/service-accounts/{id}":
    get:
      operationId: "identuum-idp-oss.service-accounts.get_service_account"
      summary: "Show a service account (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Show a service account (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
    put:
      operationId: "identuum-idp-oss.service-accounts.update_service_account"
      summary: "Update a service account (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Update a service account (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
    delete:
      operationId: "identuum-idp-oss.service-accounts.delete_service_account"
      summary: "Delete a service account (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "204":
          description: "Delete a service account (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
  "/api/v1/service-accounts/{id}/disable":
    post:
      operationId: "identuum-idp-oss.service-accounts.set_active_service_account_disable"
      summary: "Disable a service account (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "204":
          description: "Disable a service account (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
  "/api/v1/service-accounts/{id}/enable":
    post:
      operationId: "identuum-idp-oss.service-accounts.set_active_service_account_enable"
      summary: "Enable a previously-disabled service account (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "204":
          description: "Enable a previously-disabled service account (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: "identuum_session"
      description: "Browser session cookie backing a cookie-authenticated Principal (site_admin, org_admin, or any authenticated user, depending on the operation's actual role requirement — see its summary)."
    oauthClientAuth:
      type: http
      scheme: basic
      description: "OAuth 2.1 client authentication (client credentials / RFC 7592 management token) at client-management and token endpoints."
...